- Updated PC project file.
git-svn-id: https://fuji.googlecode.com/svn/trunk@277 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Source/Display.cpp
@@ -36,6 +36,6 @@ void MFDisplay_GetDisplayRect(MFRect *pRect)
 {
 	pRect->x = 0.0f;
 	pRect->y = 0.0f;
-	pRect->width = gDisplay.width;
-	pRect->height = gDisplay.height;
+	pRect->width = (float)gDisplay.width;
+	pRect->height = (float)gDisplay.height;
 }
